+++ Runtime/Core/ClickCommands/Cooldown/ITimeProvider.cs
@@ -1,0 +1,7 @@
+﻿namespace Agava.Merge2.Core
+{
+    public interface ITimeProvider
+    {
+        long NowTicks { get; }
+    }
+}