+++ CHANGELOG.md
@@ -1,9 +1,14 @@
-# com.agava.merge2 Changelog
+# Changelog
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2023-04-14
 
+### Added
+
+- `CooldownRepository` class class for storing cooldowns of clicks.
+- `CooldownCommand` class that uses `CooldownRepository` to implement a сooldown check before next commands.
 
 ## [1.1.0] - 2023-04-11